+++ cpp/CHANGELOG.md
@@ -23,6 +23,7 @@ Please mark all change in change log and use the ticket from JIRA.
 - MS-65 - Implement GetTableRowCount interface
 - MS-45 - Implement DeleteTable interface
 - MS-75 - cmake: change faiss version to 1.5.2; add CUDA gencode
+- MS-81 - fix faiss ptx issue; change cuda gencode
 
 ## Task
 - MS-74 - Change README.md in cpp